+++ src/lib_store/block_store.ml
@@ -33,8 +33,7 @@ module Block_cache =
             map_maker ~replacement:LRU ~overflow:Strong ~accounting:Precise))
        (Block_hash))
 
-(* TODO: make limits configurable *)
-let block_cache_limit = 100
+let default_block_cache_limit = 100
 
 type merge_status = Not_running | Running | Merge_failed of tztrace
 
@@ -408,22 +407,32 @@ let expected_savepoint block_store ~target_offset =
         let cycle = cemented_block_files.(nb_files - target_offset) in
         Lwt.return cycle.start_level
 
-(* [available_savepoint block_store expected_level] aims to check that
-   the [expected_level] can be used as a valid savepoint (that is to
-   say, contains metadata). It returns the [expected_level] if it is
-   valid. Returns the current savepoint otherwise .*)
-let available_savepoint block_store expected_level =
+(* [available_savepoint block_store current_head savepoint_candidate]
+   aims to check that the [savepoint_candidate] can be used as a valid
+   savepoint (that is to say, contains metadata). It returns the
+   [savepoint_candidate] block descriptor if it is valid. Returns the
+   current savepoint otherwise. *)
+let available_savepoint block_store current_head savepoint_candidate =
+  let head_hash = Block_repr.hash current_head in
   savepoint block_store >>= fun current_savepoint ->
-  if expected_level < snd current_savepoint then
-    Lwt.return (snd current_savepoint)
-  else Lwt.return expected_level
+  let new_savepoint_level =
+    if savepoint_candidate < snd current_savepoint then snd current_savepoint
+    else savepoint_candidate
+  in
+  let distance =
+    Int32.(to_int (sub (Block_repr.level current_head) new_savepoint_level))
+  in
+  (read_block ~read_metadata:false block_store (Block (head_hash, distance))
+   >>=? function
+   | Some b -> return b
+   | None -> fail (Wrong_predecessor (head_hash, distance)))
+  >>=? fun block -> return (descriptor block)
 
-(* [preserved_block block_store expected_level] returns the preserved
-   block if the given [expected_level] is higher that the current
-   preserved block. The preserved block aims to be the one needed and
-   maintained available to export snapshot. That is to say, the block:
-   lafl(head) - max_op_ttl(lafl). *)
-let preserved_block block_store current_head expected_level =
+(* [preserved_block block_store current_head] returns the
+   preserved block candidate level. The preserved block aims to be the
+   one needed and maintained available to export snapshot. That is to
+   say, the block: lafl(head) - max_op_ttl(lafl). *)
+let preserved_block block_store current_head =
   let head_hash = Block_repr.hash current_head in
   read_block_metadata block_store (Block (head_hash, 0))
   >|=? WithExceptions.Option.get ~loc:__LOC__
@@ -432,25 +441,18 @@ let preserved_block block_store current_head expected_level =
   let head_max_op_ttl =
     Int32.of_int (Block_repr.max_operations_ttl current_head_metadata)
   in
-  let block_to_preserve = Int32.(max 0l (sub head_lafl head_max_op_ttl)) in
-  let new_block_level = min block_to_preserve expected_level in
-  let distance =
-    Int32.(to_int (sub (Block_repr.level current_head) new_block_level))
-  in
-  (read_block ~read_metadata:false block_store (Block (head_hash, distance))
-   >>=? function
-   | Some b -> return b
-   | None -> fail (Wrong_predecessor (head_hash, distance)))
-  >>=? fun block -> return (descriptor block)
+  return Int32.(max 0l (sub head_lafl head_max_op_ttl))
 
 (* [infer_savepoint block_store current_head ~target_offset] returns
    the savepoint candidate for an history mode switch. *)
 let infer_savepoint block_store current_head ~target_offset =
   expected_savepoint block_store ~target_offset
   >>= fun expected_savepoint_level ->
-  available_savepoint block_store expected_savepoint_level
-  >>= fun available_savepoint ->
-  preserved_block block_store current_head available_savepoint
+  preserved_block block_store current_head >>=? fun preserved_savepoint_level ->
+  let savepoint_candidate =
+    min preserved_savepoint_level expected_savepoint_level
+  in
+  available_savepoint block_store current_head savepoint_candidate
 
 (* [expected_caboose block_store ~target_offset] computes the
    expected caboose based on the [target_offset]). None is returned if
@@ -493,7 +495,21 @@ let infer_caboose block_store savepoint current_head ~target_offset
   | Full _ -> (
       match expected_caboose block_store ~target_offset with
       | Some expected_caboose ->
-          preserved_block block_store current_head expected_caboose
+          preserved_block block_store current_head >>=? fun preserved_caboose ->
+          let new_caboose_level = min expected_caboose preserved_caboose in
+          let head_hash = Block_repr.hash current_head in
+          let distance =
+            Int32.(
+              to_int (sub (Block_repr.level current_head) new_caboose_level))
+          in
+          (read_block
+             ~read_metadata:false
+             block_store
+             (Block (head_hash, distance))
+           >>=? function
+           | Some b -> return b
+           | None -> fail (Wrong_predecessor (head_hash, distance)))
+          >>=? fun block -> return (descriptor block)
       | None -> return savepoint)
   | Rolling r ->
       let offset =
@@ -564,8 +580,8 @@ let switch_history_mode block_store ~current_head ~previous_history_mode
         (Cannot_switch_history_mode
            {previous_mode = previous_history_mode; next_mode = new_history_mode})
 
-let compute_new_savepoint block_store history_mode ~min_level_to_preserve
-    ~new_head ~cycles_to_cement =
+let compute_new_savepoint block_store history_mode ~new_store
+    ~min_level_to_preserve ~new_head ~cycles_to_cement =
   let nb_cycles_to_cement = List.length cycles_to_cement in
   assert (nb_cycles_to_cement > 0) ;
   Stored_data.get block_store.savepoint >>= fun savepoint ->
@@ -582,66 +598,78 @@ let compute_new_savepoint block_store history_mode ~min_level_to_preserve
         block_store
         ~head:new_head
         min_level_to_preserve
-      >>=? function
-      | min_block_to_preserve -> (
-          let ((_min_block_hash, min_block_level) as min_block_descr) =
-            Block_repr.descriptor min_block_to_preserve
+      >>=? fun min_block_to_preserve ->
+      let ((_min_block_hash, min_block_level) as min_block_descr) =
+        Block_repr.descriptor min_block_to_preserve
+      in
+      (* New savepoint = min min_level_to_preserve (min new lowest cemented block) *)
+      let cemented_cycles =
+        match
+          Cemented_block_store.cemented_blocks_files block_store.cemented_store
+        with
+        | None -> cycles_to_cement
+        | Some table ->
+            (Array.to_list table
+            |> List.map (fun {Cemented_block_store.start_level; end_level; _} ->
+                   (start_level, end_level)))
+            @ cycles_to_cement
+      in
+      if Compare.Int32.(snd savepoint >= min_block_level) then return savepoint
+      else
+        let cemented_cycles_len = List.length cemented_cycles in
+        (* If the offset is 0, the savepoint will be the minimum block
+           to preserve. *)
+        if offset = 0 then return min_block_descr
+        else if
+          (* If the number of cemented cycles is not yet the offset,
+             then the savepoint will be unchanged. *)
+          cemented_cycles_len < offset
+        then
+          (* In case of a freshly imported rolling snapshot, we may
+             drag the savepoint if it was not set on a cycle
+             start. Otherwise, the savepoint would be missing from the
+             store. We drag the savepoint only if it is not in the new
+             floating store nor in the cycles to cements. *)
+          let (savepoint_hash, savepoint_level) = savepoint in
+          let is_savepoint_in_cemented =
+            List.exists
+              (fun (l, h) -> l <= savepoint_level && savepoint_level <= h)
+              cycles_to_cement
           in
-          (* New savepoint = min min_level_to_preserve (min new lowest cemented block) *)
-          let cemented_cycles =
-            match
-              Cemented_block_store.cemented_blocks_files
-                block_store.cemented_store
-            with
-            | None -> cycles_to_cement
-            | Some table ->
-                (Array.to_list table
-                |> List.map
-                     (fun {Cemented_block_store.start_level; end_level; _} ->
-                       (start_level, end_level)))
-                @ cycles_to_cement
+          if not is_savepoint_in_cemented then
+            Floating_block_store.mem new_store savepoint_hash
+            >>= fun is_savepoint_in_new_store ->
+            if not is_savepoint_in_new_store then return min_block_descr
+            else return savepoint
+          else return savepoint
+        else
+          (* Else we shift the savepoint by [nb_cycles_to_cement]
+             cycles *)
+          let shifted_savepoint_level =
+            (* new lowest cemented block  *)
+            fst
+              (List.nth cemented_cycles (cemented_cycles_len - offset)
+              |> WithExceptions.Option.get ~loc:__LOC__)
           in
-          if Compare.Int32.(snd savepoint >= min_block_level) then
+          (* If the savepoint is still higher than the shifted
+             savepoint, preserve the savepoint *)
+          if Compare.Int32.(snd savepoint >= shifted_savepoint_level) then
             return savepoint
+          else if
+            (* If the new savepoint is still higher than the min block
+               to preserve, we choose the min block to preserve. *)
+            Compare.Int32.(shifted_savepoint_level >= min_block_level)
+          then return min_block_descr
           else
-            let cemented_cycles_len = List.length cemented_cycles in
-            (* If the offset is 0, the minimum block to preserve will be
-               the savepoint. *)
-            if offset = 0 then return min_block_descr
-            else if
-              (* If the number of cemented cycles is not yet the offset,
-                 then the savepoint will be unchanged. *)
-              cemented_cycles_len < offset
-            then return savepoint
-            else
-              (* Else we shift the savepoint by [nb_cycles_to_cement]
-                 cycles *)
-              let shifted_savepoint_level =
-                (* new lowest cemented block  *)
-                fst
-                  (List.nth cemented_cycles (cemented_cycles_len - offset)
-                  |> WithExceptions.Option.get ~loc:__LOC__)
-              in
-              (* If the savepoint is still higher than the shifted
-                 savepoint, preserve the savepoint *)
-              if Compare.Int32.(snd savepoint >= shifted_savepoint_level) then
-                return savepoint
-              else if
-                (* If the new savepoint is still higher than the min block
-                   to preserve, we choose the min block to preserve. *)
-                Compare.Int32.(shifted_savepoint_level >= min_block_level)
-              then return min_block_descr
-              else
-                (* Else the new savepoint is the one-cycle shifted
-                   savepoint. *)
-                read_predecessor_block_by_level_opt
-                  block_store
-                  ~head:new_head
-                  shifted_savepoint_level
-                >>=? function
-                | None ->
-                    fail (Cannot_retrieve_savepoint shifted_savepoint_level)
-                | Some savepoint -> return (Block_repr.descriptor savepoint)))
+            (* Else the new savepoint is the one-cycle shifted
+               savepoint. *)
+            read_predecessor_block_by_level_opt
+              block_store
+              ~head:new_head
+              shifted_savepoint_level
+            >>=? function
+            | None -> fail (Cannot_retrieve_savepoint shifted_savepoint_level)
+            | Some savepoint -> return (Block_repr.descriptor savepoint))
 
 let compute_new_caboose block_store history_mode ~new_savepoint
     ~min_level_to_preserve ~new_head =
@@ -683,7 +711,7 @@ module BlocksLAFL = Set.Make (Int32)
    updates the [new_store] by storing the predecessors of the
    [new_head_lafl] and preserving the
    [lowest_bound_to_preserve_in_floating]. It returns the cycles to
-   cemented from [new_head] to [cementing_highwatermark] and the
+   cement from [new_head] to [cementing_highwatermark] and the
    savepoint and caboose candidates. *)
 let update_floating_stores block_store ~history_mode ~ro_store ~rw_store
     ~new_store ~new_head ~new_head_lafl ~lowest_bound_to_preserve_in_floating
@@ -709,6 +737,8 @@ let update_floating_stores block_store ~history_mode ~ro_store ~rw_store
     final_hash
     max_nb_blocks_to_retrieve
   >>= fun lafl_predecessors ->
+  (* [min_level_to_preserve] is the lowest block that we want to keep
+     in the floating stores. *)
   let min_level_to_preserve =
     if List.length lafl_predecessors > 0 then
       Block_repr.level
@@ -792,6 +822,7 @@ let update_floating_stores block_store ~history_mode ~ro_store ~rw_store
   compute_new_savepoint
     block_store
     history_mode
+    ~new_store
     ~min_level_to_preserve
     ~new_head
     ~cycles_to_cement
@@ -878,9 +909,10 @@ let check_store_consistency block_store ~cementing_highwatermark =
               {highest_cemented_level; cementing_highwatermark}))
 
 (* We want to keep in the floating store, at least, the blocks above
-   (new_head.lafl - (new_head.lafl).max_op_ttl)) Important: we might
+   (new_head.lafl - (new_head.lafl).max_op_ttl)). Important: we might
    not have this block so it should be treated as a potential lower
-   bound. *)
+   bound. Furethermore, we consider the current caboose as a potential
+   lower bound.*)
 let compute_lowest_bound_to_preserve_in_floating block_store ~new_head
     ~new_head_metadata =
   (* Safety check: is the highwatermark consistent with our highest cemented block *)
@@ -957,7 +989,7 @@ let create_merging_thread block_store ~history_mode ~old_ro_store ~old_rw_store
           ~rw_store:old_rw_store
           ~head:new_head
       in
-      match history_mode with
+      (match history_mode with
       | History_mode.Archive ->
           List.iter_es
             (fun cycle_range ->
@@ -965,7 +997,6 @@ let create_merging_thread block_store ~history_mode ~old_ro_store ~old_rw_store
               (* In archive, we store the metadatas *)
               cement_blocks ~write_metadata:true block_store cycle)
             cycles_interval_to_cement
-          >>=? fun () -> return (new_savepoint, new_caboose)
       | Rolling offset ->
           let offset =
             (Option.value
@@ -990,9 +1021,9 @@ let create_merging_thread block_store ~history_mode ~old_ro_store ~old_rw_store
             Cemented_block_store.trigger_gc
               block_store.cemented_store
               history_mode
-            >>= fun () -> return (new_savepoint, new_caboose)
+            >>= fun () -> return_unit
           else (* Don't cement any cycles! *)
-            return (new_savepoint, new_caboose)
+            return_unit
       | Full offset ->
           let offset =
             (Option.value
@@ -1013,15 +1044,15 @@ let create_merging_thread block_store ~history_mode ~old_ro_store ~old_rw_store
             Cemented_block_store.trigger_gc
               block_store.cemented_store
               history_mode
-            >>= fun () -> return (new_savepoint, new_caboose)
+            >>= fun () -> return_unit
           else
             List.iter_es
               (fun cycle_range ->
                 cycle_reader cycle_range >>=? fun cycle ->
                 (* In full 0, we do not store the metadata *)
                 cement_blocks ~write_metadata:false block_store cycle)
-              cycles_interval_to_cement
-            >>=? fun () -> return (new_savepoint, new_caboose))
+              cycles_interval_to_cement)
+      >>=? fun () -> return (new_savepoint, new_caboose))
     (fun exn ->
       Floating_block_store.close new_ro_store >>= fun () -> Lwt.fail exn)
   >>=? fun (new_savepoint, new_caboose) ->
@@ -1049,7 +1080,6 @@ let merge_stores block_store ~(on_error : tztrace -> unit tzresult Lwt.t)
       let new_head_lafl =
         Block_repr.last_allowed_fork_level new_head_metadata
       in
-
       Store_events.(emit start_merging_stores) new_head_lafl >>= fun () ->
       check_store_consistency block_store ~cementing_highwatermark
       >>=? fun () ->
@@ -1191,7 +1221,7 @@ let may_recover_merge block_store =
               block_store.rw_floating_block_store <- rw ;
               write_status block_store Idle >>=? fun () -> return_unit))
 
-let load chain_dir ~genesis_block ~readonly =
+let load ?block_cache_limit chain_dir ~genesis_block ~readonly =
   Cemented_block_store.init chain_dir ~readonly >>=? fun cemented_store ->
   Floating_block_store.init chain_dir ~readonly RO
   >>= fun ro_floating_block_store ->
@@ -1205,7 +1235,10 @@ let load chain_dir ~genesis_block ~readonly =
   >>=? fun caboose ->
   Stored_data.init (Naming.block_store_status_file chain_dir) ~initial_data:Idle
   >>=? fun status_data ->
-  let block_cache = Block_cache.create block_cache_limit in
+  let block_cache =
+    Block_cache.create
+      (Option.value block_cache_limit ~default:default_block_cache_limit)
+  in
   let merge_scheduler = Lwt_idle_waiter.create () in
   let merge_mutex = Lwt_mutex.create () in
   let block_store =
@@ -1231,8 +1264,9 @@ let load chain_dir ~genesis_block ~readonly =
   fail_unless (status = Idle) Cannot_load_degraded_store >>=? fun () ->
   return block_store
 
-let create chain_dir ~genesis_block =
-  load chain_dir ~genesis_block ~readonly:false >>=? fun block_store ->
+let create ?block_cache_limit chain_dir ~genesis_block =
+  load chain_dir ?block_cache_limit ~genesis_block ~readonly:false
+  >>=? fun block_store ->
   store_block block_store genesis_block >>=? fun () -> return block_store
 
 let pp_merge_status fmt status =