+++ src/lib_store/test/alpha_utils.ml
@@ -386,7 +386,7 @@ let default_genesis_parameters =
     bootstrap_accounts = default_accounts;
   }
 
-let default_patch_context ctxt =
+let patch_context ctxt ~json =
   let shell =
     {
       Tezos_base.Block_header.level = 0l;
@@ -399,8 +399,6 @@ let default_patch_context ctxt =
       context = Context_hash.zero;
     }
   in
-  let open Tezos_protocol_alpha_parameters in
-  let json = Default_parameters.json_of_parameters default_genesis_parameters in
   let proto_params =
     Data_encoding.Binary.to_bytes_exn Data_encoding.json json
   in
@@ -410,6 +408,11 @@ let default_patch_context ctxt =
   Main.init ctxt shell >|= Environment.wrap_tzresult >>= function
   | Error _ -> assert false
   | Ok {context; _} -> return (Shell_context.unwrap_disk_context context)
+
+let default_patch_context ctxt =
+  patch_context
+    ctxt
+    ~json:(Default_parameters.json_of_parameters default_genesis_parameters)
 
 (********* Baking *************)